move range definition into dimension module since only exists as a dimension field
