+++ dna/integrity_zomes/sensemaker_integrity/src/cultural_context.rs
@@ -1,6 +1,6 @@
 use hdi::prelude::*;
 
-use crate::range::RangeValue;
+use crate::dimension::RangeValue;
 
 #[hdk_entry_helper]
 #[derive(Clone)]
