+++ dna/integrity_zomes/sensemaker_integrity/src/dimension.rs
@@ -1,4 +1,3 @@
-use crate::Range;
 use hdi::prelude::*;
 
 #[hdk_entry_helper]
@@ -6,4 +5,21 @@ use hdi::prelude::*;
 pub struct Dimension {
     pub name: String,
     pub range: Range,
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub struct Range {
+    pub name: String,
+    pub kind: RangeKind,
+    //not sure what `RangeValueVariant` would be used for. Leaving out for now.
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub enum RangeKind {
+    Integer { min: u32, max: u32 },
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+pub enum RangeValue {
+    Integer(u32),
 }
